CMSIS DSP Software Library now properly integrated as sub-project and compiles into a static library.
Lots of fixed-point math got upgraded to proper
floating-point, after I realized it wont cause
nearly as much computational overhead as I tought.

USB AUDIO class got a feature control unit which
handles volume and mute features. Now sample
amplitude prescaling happens in firmware.

+++ Core/Inc/usb_audio_class.h
@@ -9,7 +9,7 @@ typedef struct {
   void (*stream_start)(uint32_t sample_rate);
   void (*stream_end)(void);
   void (*audio_data_received)(int16_t* buff, uint16_t size);
-  void (*set_volume)(uint8_t volume);
+  void (*set_volume)(int16_t volume);
   void (*set_mute)(uint8_t mute);
 } USB_Audio_IntfTypeDef;
